requirements.txt
asgiref==3.11.1
awsebcli==3.27.1
bcrypt==5.0.0
blessed==1.27.0
boto3==1.42.77
botocore==1.42.77
cement==2.10.14
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
colorama==0.4.6
cryptography==46.0.6
decorator==5.2.1
Deprecated==1.3.1
Django==6.0.3
django-cors-headers==4.9.0
django-storages==1.14.6
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
fabric==3.2.2
gunicorn==25.3.0
idna==3.11
invoke==2.2.1
jmespath==1.1.0
packaging==26.0
paramiko==4.0.0
pathspec==0.12.1
pillow==12.1.1
psycopg2-binary==2.9.11
pycparser==3.0
PyJWT==2.12.1
PyNaCl==1.6.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
requests==2.33.0
s3transfer==0.16.0
semantic-version==2.10.0
setuptools==82.0.1
six==1.17.0
sqlparse==0.5.5
termcolor==2.5.0
urllib3==2.6.3
wcwidth==0.2.14
whitenoise==6.12.0
wrapt==2.1.2
python-decouple==3.8
